Modeling Planar Belt Transmissions

GitHub source

Background

Belt transmissions involve several moving elements and a variety of materials, enabling flexible power transmission but leading to complex and interdependent system design requirements. The basic performance of a candidate belt transmission can be estimated from kinematic analysis, while a dynamic simulation can give a much better idea of actual system performance and limits.

This Julia package assists users in the kinematic analysis, designing belt transmissions using flat or synchronous belts. Primary structures model pulleys and belt segments, these can be analyzed and plotted through recipes for Plots.jl. BeltTransmission was used to determine the basic design of the Moover variants.

See the package readme for release, license, and development information.

Installation

using Pkg
Pkg.add("https://github.com/mechanomy/BeltTransmission.jl.git")

Use Cases

Designing a planar belt transmission is the primary use case for BeltTransmission, but this design can have several objectives according to what is already known about the system. If the positions, sizes, and belt routing are known, these can be entered to calculate belt length and arrival and departure angles on each pulley. BeltTransmission can also be used to solve systems for other unknowns, say to determine the position of an idler pulley given a known belt length.

Basic belt transmissions

The goal of these examples is to calculate the belt length given pulley sizes and locations.

Two pulleys

using Unitful, Plots
using Geometry2D
using BeltTransmission

#create the pulleys
ctrA = Geometry2D.Point(10u"mm", 10u"mm")
pA = PlainPulley( circle=Geometry2D.Circle( center=ctrA, radius=10u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="A")
pB = PlainPulley( circle=Geometry2D.Circle(100u"mm", 100u"mm", 20u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="B")

#create the belt route
route = [pA, pB] #from A to B and back to A

#solve the departure and arrival angles on each PlainPulley
solved = calculateRouteAngles(route)

#create belt Segments between the Pulleys
segments = route2Segments(solved)
beltLength = calculateBeltLength(segments)

plot(segments, title="two pulleys")

By convention, Pulleys having a positive rotation (+uk) have a dot at their center to indicate the vector's tip, while negative rotation is shown with an X for the arrow's fletching. Once solved, the system can be interrogated:

calculateBeltLength(segments)

0.34959229994006674 m

A variety of pulleys

using Unitful, Plots
using Geometry2D
using BeltTransmission

#create the pulleys
ctrA = Geometry2D.Point(10u"mm", 10u"mm")
pA = PlainPulley( circle=Geometry2D.Circle( center=ctrA, radius=10u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="A")
pB = PlainPulley( circle=Geometry2D.Circle(100u"mm",  20u"mm", 20u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="B")
pC = PlainPulley( circle=Geometry2D.Circle( 80u"mm",  40u"mm",  5u"mm"), axis=-Geometry2D.UnitVector(0,0,1), name="C")
pD = PlainPulley( circle=Geometry2D.Circle(150u"mm",  40u"mm",  5u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="D")
pE = PlainPulley( circle=Geometry2D.Circle(100u"mm",  80u"mm",  5u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="E")

#create the belt route
route = [pA, pB, pC, pD, pE]

#solve the departure and arrival angles on each PlainPulley
solved = calculateRouteAngles(route)

#create belt Segments between the Pulleys
segments = route2Segments(solved)

plot(segments, title="Various Pulleys" )

julia> calculateBeltLength(segments)0.5507346870088579 m

julia> printSegments(segments)A:[10.000,10.000]<10.000@-90.000°>[10.000,0.000]--B:[100.000,20.000]<20.000@-90.000°>[100.000,0.000]
B:[100.000,20.000]<20.000@107.114°>[94.114,39.114]--C:[80.000,40.000]<5.000@-72.886°>[81.471,35.221]
C:[80.000,40.000]<5.000@81.787°>[80.714,44.949]--D:[150.000,40.000]<5.000@-98.213°>[149.286,35.051]
D:[150.000,40.000]<5.000@51.340°>[153.123,43.904]--E:[100.000,80.000]<5.000@51.340°>[103.123,83.904]
E:[100.000,80.000]<5.000@125.362°>[97.106,84.078]--A:[10.000,10.000]<10.000@125.362°>[4.213,18.155]
total belt length = 0.5507346870088579 m

Positive rotation of PlainPulley A will cause PlainPulley C to rotate negatively, hence the negative rotation axis on C. In addition to the PlainPulley rotation axis, the point of departure on each PlainPulley is given a slight arrow in the direction of positive belt rotation.

Multiple systems

using Unitful, Plots
using Geometry2D
using BeltTransmission

#create the pulleys
pA = PlainPulley( circle=Geometry2D.Circle( 10u"mm",  20u"mm", 20u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="B")
pB = PlainPulley( circle=Geometry2D.Circle(100u"mm",  20u"mm", 30u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="B")

pX = PlainPulley( circle=Geometry2D.Circle( 10u"mm",  20u"mm", 15u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="X")
pY = PlainPulley( circle=Geometry2D.Circle( 50u"mm",  80u"mm",  5u"mm"), axis=Geometry2D.UnitVector(0,0,1), name="Y")

#create the belt systems
segAB = route2Segments( calculateRouteAngles( [pA, pB] ))
segXY = route2Segments( calculateRouteAngles( [pX, pY] ))

plot( segAB, title="Parallel Systems", xlabel="[mm]", ylabel="[mm]", xlims=([-50,150]), ylims=([-50,100]), legend=false, dpi=100, size=(1000,500))
plot!(segXY, segmentColor=:cyan)

PlainPulley body colors are assigned by the default colormap, but segmentColor may be used to control the belt segment color. Other Plots.jl attributes may be used as expected.

Development Plan

BeltTransmission.jl is under active development, enhancement requests are welcome.

API

BeltTransmission.BeltTransmission — Module

Geometric modeling of 2D closed belt transmissions.

This module models closed belt transmissions with either:

Plain pulleys with flat belting

Timing pulleys and timing belts

With the general concept of a 'belt' entailing a continuous element, flat and timing belts add properties that affect which method may be called. For instance, a flat belt may have an arbitrary length, limited only by fabrication, while toothed belts are defined to have an integer number of teeth. Likewise, calculations involving tooth pitch are nonsensical if applied to flat or V belts. The module uses Julia's type system to differentiate between these belt types.

Timing and plain pulleys are specializaitons of AbstractPulley, permitting functions that need only the pulley location, rotation axis, or pitch diameter to accept any argument that maps to AbstractPulley, while allowing those needing tooth properties to specify arguments of TimingPulley type.

Calculations are perfomed on the pulley or belt pitch line.

Exports methods:

AbstractPulley

PlainPulley

Segment

TimingPulley

calculateBeltLength

calculateRouteAngles

calculateWrappedAngle

calculateWrappedLength

distance

findTangents

getArrivalPoint

getDeparturePoint

isSegmentMutuallyTangent

length

nGrooves2Length

nGrooves2Radius

pitchLength

printPulley

printRoute

printSegments

pulley2Circle

pulley2String

radius2NGrooves

route2Segments

toString

toStringPoints

toStringShort

toStringVectors

BeltTransmission.AbstractPulley — Type

Subtypes of AbstractPulley model a particular form of pulley. At present that includes PlainPulleys, having cylindrical faces, and TimingPulley. All subtypes have basic fields of pitch::Circle, axis::UnitVector, arrive::Angle, depart::Angle, and a name::String.

BeltTransmission.PlainPulley — Type

Models a cylindrical plain pulley in a BeltTransmission with:

pitch

the pitch Circle

axis

the rotation axis for the pulley with +/- defining the 'positive' rotation direction

arrive

angle of the radial vector of the belt's point of arrival

depart

angle of the radial vector of the belt's point of departure

name

convenience name of the pulley

BeltTransmission.PlainPulley — Method

PlainPulley(circle::Geometry2D.Circle, axis::Geometry2D.UnitVector, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission, described by a circle, rotation axis, and name.

BeltTransmission.PlainPulley — Method

PlainPulley(center::Geometry2D.Point, radius::Unitful.Length, axis::Geometry2D.UnitVector, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission, having a pitch diameter, rotation axis, angles arrive and depart when rotated postively according to the axis, and an optional name.

BeltTransmission.PlainPulley — Method

PlainPulley(center::Geometry2D.Point, radius::Unitful.Length, axis::Geometry2D.UnitVector) :: PlainPulley

Models a PlainPulley in a BeltTransmission, having a pitch diameter, rotation axis, angles arrive and depart when rotated postively according to the axis, and an optional name.

BeltTransmission.PlainPulley — Method

PlainPulley(center::Geometry2D.Point, radius::Unitful.Length, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission, located at center with pitch radius and name.

BeltTransmission.PlainPulley — Method

PlainPulley(center::Geometry2D.Point, radius::Unitful.Length) :: PlainPulley

Models a PlainPulley in a BeltTransmission, located at center with pitch radius.

BeltTransmission.Segment — Type

Describes a belt segment between depart and arrive Pulleys.

depart

Departing PlainPulley

arrive

Arriving PlainPulley

BeltTransmission.TimingPulley — Type

Models a TimingPulley in a BeltTransmission, described by a pitch circle, rotation axis and beltPitch.

Base.length — Method

Base.length(seg::Segment) :: Unitful.Length

Returns the straight-line distance or length of Segment seg via distance.

Base.show — Method

Base.show(io::IO, p::AbstractPulley)

Function to show() a AbstractPulley via pulley2String.

Base.show — Method

Base.show(io::IO, seg::Segment)

show()s the Segment via toStringShort.

BeltTransmission.calculateBeltLength — Method

calculateBeltLength(route::Vector{PlainPulley}) :: Unitful.Length

Calculates the belt length over the given route as the sum of circular sections at the PlainPulley pitch radii between the arrival and departure angles.

BeltTransmission.calculateBeltLength — Method

calculateBeltLength(segments::Vector{Segment}) :: Unitful.Length

Calculates the belt length over the given route as the sum of straight Segments.

BeltTransmission.calculateRouteAngles — Method

calculateRouteAngles(route::Vector{PlainPulley}, plotSegments::Bool=false) :: Vector{PlainPulley}

Given an ordered vector of Pulleys, output a vector of new Pulleys whose arrive and depart angles are set to connect the pulleys with mutually tangent segments. Convention: pulleys are listed in 'positive' belt rotation order, consistent with the direction of each PlainPulley's rotation axis.

BeltTransmission.calculateWrappedAngle — Method

calculateWrappedAngle(p::AbstractPulley) :: Geometry2D.Angle

Given p, calculate the wrapped angle from p.arrive to p.depart. Note that the wrapped angle is not restricted to <= 1 revolution, as the pulley may be wrapped multiple times.

BeltTransmission.calculateWrappedLength — Method

calculateWrappedLength(p::AbstractPulley) :: Unitful.Length

Given p, calculate the arclength of the wrapped segment from p.arrive to p.depart Note that the wrapped length is not restricted to <= 1 revolution, as the pulley may be wrapped multiple times.

BeltTransmission.distance — Method

distance(seg::Segment) :: Unitful.Length

Returns the straight-line distance or length of Segment seg.

BeltTransmission.findTangents — Method

findTangents(; a::PlainPulley, b::PlainPulley, plotResult::Bool=false) :: Vector{Segment}

Find four lines tangent to both PlainPulley a and b, returns 4 Segments with departure and arrival angles locating the points of tangency on the circles.

BeltTransmission.getArrivalPoint — Method

getArrivalPoint(p::AbstractPulley)::Geometry2D.Point

Returns the point of arrival.

BeltTransmission.getArrivalPoint — Method

getArrivalPoint(seg::Segment) :: Geometry2D.Point

Returns the arrival Geometry2D.Point of seg.

BeltTransmission.getDeparturePoint — Method

getDeparturePoint(p::AbstractPulley)::Geometry2D.Point

Returns the point of departure.

BeltTransmission.getDeparturePoint — Method

getDeparturePoint(seg::Segment) :: Geometry2D.Point

Returns the departure Geometry2D.Point of seg.

BeltTransmission.isSegmentMutuallyTangent — Method

isSegmentMutuallyTangent( seg::Segment ) :: Bool

Determines whether the departure and arrival points are tangent to the connecting belt segment.

For a segment a->b between ai&bi; the correct arrival and departure angles will have cross products that match both axes, (ra-a1)x(a1-b1) = ax1 && (rb-b1)x(a1-b1) = bx1, all others should be false.

BeltTransmission.nGrooves2Length — Method

nGrooves2Length(pitch::Unitful.Length, nGrooves::Integer)::Unitful.Length

Convert the pitch and nGrooves to the circumferential length.

BeltTransmission.nGrooves2Radius — Method

nGrooves2Radius(pitch::Unitful.Length, nGrooves::Integer)::Unitful.Length

Convert the pitch and nGrooves to the pitch radius.

BeltTransmission.pitchLength — Method

pitchLength(p::AbstractPulley) :: Unitful.Length

Returns the circumferential length of the pitch diameter of the pulley.

BeltTransmission.printPulley — Method

printPulley(p::AbstractPulley)

Prints the result of pulley2String to the standard output

BeltTransmission.printRoute — Method

printRoute(route::Vector{PlainPulley})

Prints the Pulleys and total belt length for the given route.

BeltTransmission.printSegments — Method

printSegments(segments::Vector{Segment})

Prints the Segments, and total belt length for the given route.

BeltTransmission.pulley2Circle — Method

pulley2Circle(p::AbstractPulley) :: Geometry2D.Circle

Returns the pitch Circle of p.

BeltTransmission.pulley2String — Method

pulley2String(p::PlainPulley) :: String

Returns a descriptive string of the given PlainPulley p of the form: pulley[struct] @ [1.000,2.000] r[3.000] mm arrive[57.296°] depart[114.592°] aWrap[57.296°] lWrap[3.000]"

BeltTransmission.pulley2String — Method

pulley2String(p::TimingPulley) :: String

Returns a descriptive string of the given TimingPulley p of the form: timing pulley[struct] @ [1.000,2.000] r[3.000] mm arrive[57.296°] depart[114.592°] aWrap[57.296°] lWrap[3.000]"

BeltTransmission.radius2NGrooves — Method

radius2NGrooves(pitch::Unitful.Length, radius::Unitful.Length)::Integer

Convert the pitch and radius to the number of grooves.

BeltTransmission.route2Segments — Method

route2Segments(route::Vector{PlainPulley}) :: Vector{Segment}

Given the ordered Pulleys of a belt routing, returns a vector of the free-space Segments connecting the Pulleys.

BeltTransmission.toString — Method

toString(seg::Segment) :: String

Calls toStringShort.

BeltTransmission.toStringPoints — Method

toStringPoints(seg::Segment) :: String

Creates strings like: Segment: depart[100.000, 110.000] – arrive[-100.000, 110.000] length[200.000]

BeltTransmission.toStringShort — Method

toStringShort(seg::Segment) :: String

Returns a short string of the form 'A – B', for a departing PlainPulley named A arriving at a PlainPulley named B.

BeltTransmission.toStringVectors — Method

toStringVectors(seg::Segment) :: String

Creates strings like: "A:[100.000,100.000]<10.000@90.000°>[100.000,110.000]–B:[-100.000,100.000]<10.000@90.000°>[-100.000,110.000]"

KeywordDispatch.kwcall — Method

PlainPulley(; center::Geometry2D.Point, radius::Unitful.Length, axis::Geometry2D.UnitVector, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission through keyword arguments.

KeywordDispatch.kwcall — Method

PlainPulley(; circle::Geometry2D.Circle, axis::Geometry2D.UnitVector, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission through keyword arguments.

KeywordDispatch.kwcall — Method

PlainPulley(; circle::Geometry2D.Circle, axis::Geometry2D.UnitVector, arrive::Geometry2D.Radian, depart::Geometry2D.Radian, name::String) :: PlainPulley

Models a PlainPulley in a BeltTransmission through keyword arguments.

KeywordDispatch.kwcall — Method

TimingPulley(; pitch::Geometry2D.Circle, axis::Geometry2D.UnitVector, beltPitch::Unitful.Length, arrive::Geometry2D.Radian, depart::Geometry2D.Radian, name::String) :: TimingPulley

Models a TimingPulley in a BeltTransmission, described by a pitch circle, rotation axis and beltPitch. Optional angles arrive and depart specify belt's arrival and departure tangent points. name identifies the pulley when plotting and printing.

KeywordDispatch.kwcall — Method

TimingPulley(; center::Geometry2D.Point, radius::Unitful.Length, axis::Geometry2D.UnitVector, beltPitch::Unitful.Length, arrive::Geometry2D.Radian, depart::Geometry2D.Radian, name::String)

Models a TimingPulley in a BeltTransmission, described by a circle, rotation axis, and name.

KeywordDispatch.kwcall — Method

TimingPulley(; center::Geometry2D.Point, nGrooves::Integer, axis::Geometry2D.UnitVector, beltPitch::Unitful.Length, name::String)

Models a TimingPulley in a BeltTransmission, described by a circle, rotation axis, and name. nGrooves and beltPitch are used to find the pulley pitch diameter.

KeywordDispatch.kwcall — Method

Segment(; depart::PlainPulley, arrive::PlainPulley) :: Segment

Create a belt Segment between depart and arrive Pulleys

RecipesBase.apply_recipe — Method

plotRecipe(p::PlainPulley; n=100, lengthUnit=u"mm", segmentColor=:magenta, arrowFactor=0.03)

A plot recipe for plotting Pulleys under Plots.jl. Keyword n can be used to increase the number of points constituting the pulley edge. lengthUnit is a Unitful unit for scaling the linear axes. arrowFactor controls the size of the arrow head at depart.

using Plots, Unitful, BeltTransmission, Geometry2D
p = PlainPulley( Geometry2D.Circle(1u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
plot(p)

RecipesBase.apply_recipe — Method

plotRecipe(seg::Segment; n=100, lengthUnit=u"mm", segmentColor=:magenta, arrowFactor=0.03)

Plot recipe to plot the free sections of a segment, does not plot the pulleys.

using Plots, Unitful, BeltTransmission, Geometry2D
a = PlainPulley( Geometry2D.Circle(1u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
b = PlainPulley( Geometry2D.Circle(10u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
seg = Segment(depart=a, arrive=b)
plot(seg)

RecipesBase.apply_recipe — Method

plotRecipe(p::TimingPulley; n=100, lengthUnit=u"mm", segmentColor=:magenta, arrowFactor=0.03)

A plot recipe for plotting Pulleys under Plots.jl. Keyword n can be used to increase the number of points constituting the pulley edge. lengthUnit is a Unitful unit for scaling the linear axes. arrowFactor controls the size of the arrow head at depart.

using Plots, Unitful, BeltTransmission, Geometry2D
p = TimingPulley( Geometry2D.Circle(1u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
plot(p)

RecipesBase.apply_recipe — Method

plotRecipe(route::Vector{PlainPulley})

Plots the Pulleys in a route.

using Plots, Unitful, BeltTransmission, Geometry2D
a = PlainPulley( Geometry2D.Circle(1u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
b = PlainPulley( Geometry2D.Circle(10u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
route = calculateRouteAngles([a,b])
plot(route)

RecipesBase.apply_recipe — Method

plotRecipe(segments::Vector{Segment})

Plots the Pulleys and Segments in a route.

using Plots, Unitful, BeltTransmission, Geometry2D
a = PlainPulley( Geometry2D.Circle(1u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
b = PlainPulley( Geometry2D.Circle(10u"mm",2u"mm",3u"mm"), Geometry2D.uk, "recipe" )
route = calculateRouteAngles([a,b])
segments = route2Segments(route)
plot(segments)